+++ test/dart_api_test.dart
@@ -185,4 +185,44 @@ main() {
       expect(css, equals("a {\n  b: from-importer;\n}"));
     });
   });
+
+  group("charset", () {
+    group("= true", () {
+      test("doesn't emit @charset for a pure-ASCII stylesheet", () {
+        expect(compileString("a {b: c}"), equals("""
+a {
+  b: c;
+}"""));
+      });
+
+      test("emits @charset with expanded output", () async {
+        expect(compileString("a {b: 👭}"), equals("""
+@charset "UTF-8";
+a {
+  b: 👭;
+}"""));
+      });
+
+      test("emits a BOM with compressed output", () async {
+        expect(compileString("a {b: 👭}", style: OutputStyle.compressed),
+            equals("\u{FEFF}a{b:👭}"));
+      });
+    });
+
+    group("= false", () {
+      test("doesn't emit @charset with expanded output", () async {
+        expect(compileString("a {b: 👭}", charset: false), equals("""
+a {
+  b: 👭;
+}"""));
+      });
+
+      test("emits a BOM with compressed output", () async {
+        expect(
+            compileString("a {b: 👭}",
+                charset: false, style: OutputStyle.compressed),
+            equals("a{b:👭}"));
+      });
+    });
+  });
 }